requirements.txt
# requirements.txt

streamlit
openai
pillow
requests
tweepy
atro_python
python-dotenv
atproto